[WIP] CMC shutter page
